backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-coach-backend"
version = "0.1.0"
description = "AI Coach Training Platform - Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.5.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.1",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "websockets>=13.0",
    "sse-starlette>=2.0.0",
    "openai>=1.50.0",
    "anthropic>=0.40.0",
    "pypdf>=5.0.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    "aiofiles>=23.0.0",
    "azure-identity>=1.17.0",
    "azure-ai-projects>=2.0.1",  # Must use v2.0+ Agent Registry API (create_version, Responses API)
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-timeout>=2.2.0",
    "pytest-cov>=5.0.0",
    "pytest-html>=4.1.0",
    "ruff>=0.6.0",
]
postgresql = [
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
]
voice = [
    "azure-cognitiveservices-speech>=1.48.0",
    "azure-ai-voicelive[aiohttp]>=1.0.0",
]
all = [
    "ai-coach-backend[dev,postgresql,voice]",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
timeout = 60
addopts = "-m 'not integration'"
markers = [
    "integration: mark test as requiring real Azure credentials (deselect with '-m \"not integration\"')",
]

[tool.ruff]
target-version = "py311"
line-length = 100
exclude = ["alembic/"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
